+++ tests/integration/test_dataset/test_dataset.py
@@ -10,6 +10,7 @@ from tests.common_test_utils import (
     is_evidence_present,
     create_output_dir,
     assert_no_errors,
+    close_test_redis_server,
     get_slips_test_command,
     skip_if_missing_runtime_dependencies,
 )
@@ -72,26 +73,31 @@ def test_binetflow(
         python_modules=("termcolor",), binaries=("redis-server",)
     )
     output_dir = create_output_dir(output_dir)
+    success = False
+    try:
+        output_file = os.path.join(output_dir, "slips_output.txt")
+        command = get_slips_test_command(
+            f"-e 1 -t -o {output_dir} -P {redis_port} -f {binetflow_path}"
+        )
+        command = f"{command} > {output_file} 2>&1"
+        # this function returns when slips is done
+        run_slips(command)
 
-    output_file = os.path.join(output_dir, "slips_output.txt")
-    command = get_slips_test_command(
-        f"-e 1 -t -o {output_dir} -P {redis_port} -f {binetflow_path}"
-    )
-    command = f"{command} > {output_file} 2>&1"
-    # this function returns when slips is done
-    run_slips(command)
+        assert_no_errors(output_dir)
 
-    assert_no_errors(output_dir)
+        database = ModuleFactory().create_db_manager_obj(
+            redis_port, output_dir=output_dir, start_redis_server=False
+        )
+        profiles = database.get_profiles_len()
+        assert profiles > expected_profiles
 
-    database = ModuleFactory().create_db_manager_obj(
-        redis_port, output_dir=output_dir, start_redis_server=False
-    )
-    profiles = database.get_profiles_len()
-    assert profiles > expected_profiles
-
-    log_file = output_dir / "alerts" / alerts_file
-    assert is_evidence_present(log_file, expected_evidence) is True
-    shutil.rmtree(output_dir)
+        log_file = output_dir / "alerts" / alerts_file
+        assert is_evidence_present(log_file, expected_evidence) is True
+        success = True
+    finally:
+        if success:
+            close_test_redis_server(redis_port)
+            shutil.rmtree(output_dir)
 
 
 @pytest.mark.parametrize(
@@ -114,27 +120,35 @@ def test_suricata(suricata_path, output_dir, redis_port, expected_evidence):
         python_modules=("termcolor",), binaries=("redis-server",)
     )
     output_dir = create_output_dir(output_dir)
-    output_file = os.path.join(output_dir, "slips_output.txt")
-    command = get_slips_test_command(
-        f"-e 1 -t -f {suricata_path} -o {output_dir} -P {redis_port}"
-    )
-    command = f"{command} > {output_file} 2>&1"
-    # this function returns when slips is done
-    run_slips(command)
+    success = False
+    try:
+        output_file = os.path.join(output_dir, "slips_output.txt")
+        command = get_slips_test_command(
+            f"-e 1 -t -f {suricata_path} -o {output_dir} -P {redis_port}"
+        )
+        command = f"{command} > {output_file} 2>&1"
+        # this function returns when slips is done
+        run_slips(command)
 
-    assert_no_errors(output_dir)
+        assert_no_errors(output_dir)
 
-    database = ModuleFactory().create_db_manager_obj(
-        redis_port, output_dir=output_dir, start_redis_server=False
-    )
-    profiles = database.get_profiles_len()
-    # todo the profiles should be way more than 10, maybe 76, but it varies
-    #  each run, we need to sy why
-    assert profiles > 10
+        database = ModuleFactory().create_db_manager_obj(
+            redis_port, output_dir=output_dir, start_redis_server=False
+        )
+        profiles = database.get_profiles_len()
+        # todo the profiles should be way more than 10, maybe 76, but it varies
+        #  each run, we need to sy why
+        assert profiles > 10
 
-    log_file = output_dir / "alerts" / alerts_file
-    assert any(is_evidence_present(log_file, ev) for ev in expected_evidence)
-    shutil.rmtree(output_dir)
+        log_file = output_dir / "alerts" / alerts_file
+        assert any(
+            is_evidence_present(log_file, ev) for ev in expected_evidence
+        )
+        success = True
+    finally:
+        if success:
+            close_test_redis_server(redis_port)
+            shutil.rmtree(output_dir)
 
 
 @pytest.mark.skipif(
@@ -153,24 +167,26 @@ def test_nfdump(nfdump_path, output_dir, redis_port):
         python_modules=("termcolor",), binaries=("redis-server",)
     )
     output_dir = create_output_dir(output_dir)
+    success = False
+    try:
+        # expected_evidence = 'Connection to unknown destination port 902/TCP'
 
-    # expected_evidence = 'Connection to unknown destination port 902/TCP'
+        output_file = os.path.join(output_dir, "slips_output.txt")
+        command = get_slips_test_command(
+            f"-e 1 -t -f {nfdump_path} -o {output_dir} -P {redis_port}"
+        )
+        command = f"{command} > {output_file} 2>&1"
+        # this function returns when slips is done
+        run_slips(command)
 
-    output_file = os.path.join(output_dir, "slips_output.txt")
-    command = get_slips_test_command(
-        f"-e 1 -t -f {nfdump_path} -o {output_dir} -P {redis_port}"
-    )
-    command = f"{command} > {output_file} 2>&1"
-    # this function returns when slips is done
-    run_slips(command)
-
-    database = ModuleFactory().create_db_manager_obj(
-        redis_port, output_dir=output_dir, start_redis_server=False
-    )
-    profiles = database.get_profiles_len()
-    assert_no_errors(output_dir)
-    assert profiles > 0
-
-    # log_file = os.path.join(output_dir, alerts_file)
-    # assert is_evidence_present(log_file, expected_evidence) == True
-    shutil.rmtree(output_dir)
+        database = ModuleFactory().create_db_manager_obj(
+            redis_port, output_dir=output_dir, start_redis_server=False
+        )
+        profiles = database.get_profiles_len()
+        assert_no_errors(output_dir)
+        assert profiles > 0
+        success = True
+    finally:
+        if success:
+            close_test_redis_server(redis_port)
+            shutil.rmtree(output_dir)
